Problem #52
I wish the problem of the day was flipped with yesterdays (#51), as this was literally something I had as a step to the final version of #51's solution. While thinking through the logic, it was easier for me to separate out the board management and simply count the number of valid solutions, then re-implement it later.